one folder for one paper

+++ utils.py
@@ -10,11 +10,12 @@ import dendropy
 
 MAX_SIZE = 1024 * 1024 * 10
 
-NEXUS_SUFFIX = '.nex,.nexus'.split(',')
+NEXUS_SUFFIX = set('.nex,.nexus'.split(','))
 TREE_SUFFIX = set('.nwk,.newick,.nex,.nexus,.tre,.tree,.treefile'.split(','))
-TXT_SUFFIX = set(['.txt'])
-ZIP_SUFFIX = set(['.zip'])
-OUT_FOLDER = Path(r'R:\dryad_out')
+TXT_SUFFIX = {'.txt'}
+ZIP_SUFFIX = {'.zip'}
+TARGET_SUFFIX = TREE_SUFFIX | ZIP_SUFFIX | TXT_SUFFIX | NEXUS_SUFFIX
+OUT_FOLDER = Path(r'R:\tree_crawl_out').absolute()
 if not OUT_FOLDER.exists():
     OUT_FOLDER.mkdir()
 
@@ -32,7 +33,7 @@ class Result:
     pub_date: str = ''
     title: str = ''
     volume: int = 0
-    tree_files: tuple = tuple()
+    tree_files: tuple[Path] = tuple()
 
     def empty(self):
         return len(self.tree_files) == 0
@@ -63,6 +64,8 @@ def get_doi(raw_doi: str, doi_type='default') -> str:
         return f'doi:{doi}'
     elif doi_type == 'dryad':
         return f'doi%3A{doi.replace("/", "%2F")}'
+    elif doi_type == 'folder':
+        return doi.replace('/', '_')
     else:
         return doi
 
@@ -99,35 +102,39 @@ def is_valid_tree(tmpfile: Path) -> bool:
     return False
 
 
-def extract_tree(z: ZipFile):
-    # todo: no valid tree?
+def extract_tree(z: ZipFile, out_folder: Path):
+    # since one paper one folder, it's hardly to overwrite existed file
     for file in z.namelist():
         suffix = Path(file).suffix.lower()
+        if suffix not in TARGET_SUFFIX:
+            print('\t', file, 'is not tree file')
+            continue
+        tmpfile = out_folder / file
+        z.extract(file, path=out_folder)
         if suffix in TXT_SUFFIX or suffix in NEXUS_SUFFIX:
-            z.extract(file, path=OUT_FOLDER)
-            tmpfile = OUT_FOLDER / file
             if is_valid_tree(tmpfile):
-                yield file
+                yield tmpfile
             else:
                 print('\t', file, 'is not tree file')
                 tmpfile.unlink()
         elif suffix in TREE_SUFFIX:
-            z.extract(file, path=OUT_FOLDER)
-            yield file
+            z.extract(file, path=out_folder)
+            yield tmpfile
         elif suffix in ZIP_SUFFIX:
             print('\t', 'Extracting', file)
-            z.extract(file, path=OUT_FOLDER)
-            with ZipFile(OUT_FOLDER / file, 'r') as zz:
-                yield from extract_tree(zz)
+            z.extract(file, path=out_folder)
+            with ZipFile(out_folder / file, 'r') as zz:
+                yield from extract_tree(zz, out_folder)
+            tmpfile.unlink()
         else:
-            print('\t', file, 'is not tree file')
+            pass
 
 
-def filter_tree_from_zip(file_bin: bytes) -> list:
+def filter_tree_from_zip(file_bin: bytes, out_folder: Path) -> list:
     # filter tree files from zip
     # extract trees into OUT_FOLDER/
     tree_files = list()
     with ZipFile(BytesIO(file_bin), 'r') as z:
-        for tree_file in extract_tree(z):
+        for tree_file in extract_tree(z, out_folder):
             tree_files.append(tree_file)
     return tree_files
